+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "datarobot-genai"
-version = "0.2.2"
+version = "0.2.1"
 description = "Generic helpers for GenAI"
 readme = "README.md"
 requires-python = ">=3.10, <3.13"
@@ -36,7 +36,7 @@ datarobot_llm_clients = "datarobot_genai.nat.datarobot_llm_clients"
 
 [project.optional-dependencies]
 crewai = [
-  "crewai>=0.193.2,<1.0.0",
+  "crewai>=1.1.0",
   "crewai-tools[mcp]>=0.69.0,<0.77.0",
   "opentelemetry-instrumentation-crewai>=0.40.5,<1.0.0",
   "pybase64>=1.4.2,<2.0.0",
@@ -60,8 +60,8 @@ llamaindex = [
 nat = [
   "nvidia-nat==1.3.0; python_version >= '3.11'",
   "nvidia-nat-opentelemetry==1.3.0; python_version >= '3.11'",
-  "nvidia-nat-crewai==1.3.0; python_version >= '3.11'",
   "nvidia-nat-langchain==1.3.0; python_version >= '3.11'",
+  "crewai>=1.1.0; python_version >= '3.11'",
   "llama-index-llms-litellm>=0.4.1,<0.7.0",  # Need this to support datarobot-llm plugin
   "opentelemetry-instrumentation-crewai>=0.40.5,<1.0.0",
   "opentelemetry-instrumentation-llamaindex>=0.40.5,<1.0.0",
@@ -85,7 +85,7 @@ drmcp = [
   "aiohttp>=3.9.0,<4.0.0",
   "aiohttp-retry>=2.8.3,<3.0.0",
   "aiosignal>=1.3.1,<2.0.0",
-  "fastmcp==2.13.0.2",
+  "fastmcp>=2.13.0.2,<3.0.0",
 ]
 
 [tool.hatch.build.targets.wheel]
@@ -106,6 +106,9 @@ include = [
 package = true
 override-dependencies = [
   "ragas>=0.3.8",
+]
+exclude-dependencies = [
+  "uv",
 ]
 
 [dependency-groups]
